backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.34
pymysql==1.1.1
cryptography==43.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
redis==5.0.8
structlog==24.4.0
python-multipart==0.0.10
onvif-zeep==0.2.12
av==12.3.0
httpx==0.27.2
